fix : bunch of bug fixes discovered when pairing with the engine

+++ src/ast/utils/numbers.rs
@@ -66,19 +66,19 @@ fn float(input: Span) -> IResult<Span, Span> {
     ))(input)
 }
 
-pub fn parse_number(input: Span) -> PResult<u64> {
+pub fn parse_number(input: Span) -> PResult<u128> {
     alt((
         map_res(octal, |value| {
-            u64::from_str_radix(&value.replace("_", ""), 8)
+            u128::from_str_radix(&value.replace("_", ""), 8)
         }),
         map_res(hexadecimal, |value| {
-            u64::from_str_radix(&value.replace("_", ""), 16)
+            u128::from_str_radix(&value.replace("_", ""), 16)
         }),
         map_res(binary, |value| {
-            u64::from_str_radix(&value.replace("_", ""), 2)
+            u128::from_str_radix(&value.replace("_", ""), 2)
         }),
         map_res(decimal, |value| {
-            u64::from_str_radix(&value.replace("_", ""), 10)
+            u128::from_str_radix(&value.replace("_", ""), 10)
         }),
     ))(input)
 }
